requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
dnspython==2.5.0
python-whois==0.8.0
beautifulsoup4==4.12.3
spacy==3.7.2
email-validator==2.1.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
